26-ый commit 14-ый спринт, домашняя работа, исправление замечаний
