+++ app/src/main/java/com/example/playlistmaker/App.kt
@@ -6,7 +6,7 @@ import androidx.appcompat.app.AppCompatDelegate
 class App : Application() {
 
     var darkTheme = false
-
+    var debounce = Debounce
     override fun onCreate() {
         super.onCreate()
         val darkTheme = DarkTheme(getSharedPreferences(DarkTheme.DARK_THEME, MODE_PRIVATE))
