+++ app/src/main/java/com/example/playlistmaker/Debounce.kt
@@ -1,0 +1,28 @@
+package com.example.playlistmaker
+
+import android.os.Handler
+import android.os.Looper
+
+object Debounce {
+
+
+    private const val CLICK_DEBOUNCE_DELAY = 1000L
+    private const val SEARCH_DEBOUNCE_DELAY = 2000L
+    private var isClickAllowed = true
+    private val handler = Handler(Looper.getMainLooper())
+
+    fun clickDebounce(): Boolean {
+        val current = isClickAllowed
+        if (isClickAllowed) {
+            isClickAllowed = false
+            handler.postDelayed({ isClickAllowed = true }, CLICK_DEBOUNCE_DELAY)
+        }
+        return current
+    }
+
+    fun debounce(runnable: Runnable) {
+        handler.removeCallbacks(runnable)
+        handler.postDelayed(runnable, SEARCH_DEBOUNCE_DELAY)
+    }
+
+}
